requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-openai>=0.1
openai>=1.0
httpx>=0.27
arize-otel
openinference-instrumentation-langchain
python-dotenv>=1.0
ipywidgets>=8.0
